+++ packages/terra-clinical-item-display/src/terra-dev-site/doc/example/DefaultComment.jsx
@@ -5,7 +5,6 @@ import ItemComment from 'terra-clinical-item-display/lib/ItemComment';
 const component = () => (
   <span>
     <ItemComment text="Comment" />
-    <ItemComment text="Disabled Comment" isDisabled />
   </span>
 );