+++ packages/terra-clinical-item-display/src/terra-dev-site/doc/example/TextStylesDisabled.jsx
@@ -1,3 +1,5 @@
+// TODO: remove this example in the next major version.
+
 import React from 'react';
 import ItemDisplay from 'terra-clinical-item-display';